bug fix on `SkipMCPCheckChallenges`
As long as PhoenixBuilder can log on to the server and the Minecraft packet protocol is correct, it can run on any Netease version of the rental server such as Netease 2.11.

However, it should be noted that this is only an alternative solution to ease the trouble caused by NetEase updates.